Refactored of Sionna server script in a pythonic way and started working on LTE-V2X interference
